pyproject.toml
[tool.poetry]
name = "datapipe-label-studio-lite"
version = "0.4.0"
description = ""
authors = [
    "Alexander Kozlov <bobokvsky@epoch8.co>",
    "Andrey Tatarinov <a@tatarinov.co>",
]

exclude = ["tests", "experiments"]

[tool.poetry.dependencies]
python = ">=3.8, <3.12"
pandas = ">=1.2.0"
numpy = ">=1.21.0"
SQLAlchemy = ">=1.4.25"
requests = ">=2.24.0"
datapipe-core = ">=0.14.0, <0.15"
label-studio-sdk = ">=0.0.34"

[tool.poetry.dev-dependencies]
flake8 = "*"
black = "*"

datapipe-core = {version=">=0.13.4, <0.14", extras=["sqlite"]}

pytest = "*"
pytest-xdist = "*"
pytest-cases = "^3.6.4"

mypy = "*"
pandas-stubs = "*"
types-requests = "*"
types-setuptools = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
